Migrate to webpack
