+++ .gitignore
@@ -1,4 +1,5 @@
 out
+dist
 node_modules
 npm-debug.log
 .vscode-test
