+++ .vscode/settings.json
@@ -1,13 +1,11 @@
-// Place your settings in this file to overwrite default and user settings.
 {
-  "editor.tabSize": 2,
-  "editor.insertSpaces": true,
-  /*"files.exclude": {
-    "node_modules": true,
-    "out": false // set this to true to hide the "out" folder with the compiled JS files
-  },
-  "search.exclude": {
-    "out": true // set this to false to include "out" folder in search results
-  },*/
-  "typescript.tsdk": "./node_modules/typescript/lib" // we want to use the TS server from our node_modules folder to control its version
+    "editor.insertSpaces": true,
+    "editor.formatOnSave": true,
+    "files.exclude": {
+        "out": false
+    },
+    "search.exclude": {
+        "out": true
+    },
+    "typescript.tsdk": "./node_modules/typescript/lib"
 }
