+++ package.json
@@ -9,7 +9,7 @@
   "displayName": "CodeMetrics",
   "icon": "images/logo.png",
   "engines": {
-    "vscode": "^1.16.0"
+    "vscode": "^1.37.0"
   },
   "repository": {
     "type": "git",
@@ -1161,23 +1161,29 @@
       }
     }
   },
-  "main": "./out/src/extension",
+  "main": "./dist/extension",
   "scripts": {
-    "vscode:prepublish": "tsc -p ./",
-    "compile": "tsc -watch -p ./",
-    "postinstall": "node ./node_modules/vscode/bin/install"
+    "vscode:prepublish": "webpack --mode production",
+    "webpack": "webpack --mode development",
+    "webpack-prod": "webpack --mode production",
+    "webpack-dev": "webpack --mode development --watch",
+    "analyze": "webpack --mode development --json | npx webpack-bundle-size-analyzer"
   },
   "devDependencies": {
-    "@types/minimatch": "^2.0.29",
-    "@types/node": "^7.10.6",
-    "vscode": "^1.1.34"
+    "@types/minimatch": "^3.0.3",
+    "@types/node": "^12.6.8",
+    "ts-loader": "^6.0.4",
+    "tsconfig-paths": "^3.8.0",
+    "vscode": "^1.1.35",
+    "webpack": "^4.36.1",
+    "webpack-cli": "^3.3.6"
   },
   "dependencies": {
     "luaparse": "^0.2.1",
     "minimatch": "^3.0.4",
     "tsmetrics-core": "^1.0.2",
-    "typescript": "^2.9.2",
-    "vscode-languageclient": "^3.5.1",
-    "vscode-languageserver": "^3.5.1"
+    "typescript": "^3.5.3",
+    "vscode-languageclient": "5.3.0-next.6",
+    "vscode-languageserver": "5.3.0-next.7"
   }
 }
